+++ tests/test_graph.py
@@ -1,9 +1,10 @@
 """
 Tests for the graph object.
 """
-import scipy
+import scipy as sp
 import scipy.sparse
 import math
+import pytest
 from context import sgtl
 
 # Define the adjacency matrices of some useful graphs.
@@ -54,3 +55,89 @@ def test_graph_constructor():
     assert graph.num_edges == 21
     assert graph.degrees[2] == 4
     assert graph.degrees[4] == 5
+
+
+def test_complete_graph():
+    # Create a complete graph
+    n = 4
+    graph = sgtl.graph.complete_graph(n)
+    expected_adjacency_matrix = sp.sparse.csr_matrix([[0, 1, 1, 1], [1, 0, 1, 1], [1, 1, 0, 1], [1, 1, 1, 0]])
+
+    assert graph.num_vertices == 4
+    adj_mat_diff = (graph.adj_mat - expected_adjacency_matrix)
+    adj_mat_diff.eliminate_zeros()
+    assert adj_mat_diff.nnz == 0
+
+    # Make sure we can't do something stupid
+    with pytest.raises(ValueError):
+        _ = sgtl.graph.complete_graph(-10)
+    with pytest.raises(ValueError):
+        _ = sgtl.graph.complete_graph(0)
+
+
+def test_cycle_graph():
+    # Create a cycle graph
+    n = 5
+    graph = sgtl.graph.cycle_graph(n)
+    expected_adjacency_matrix = sp.sparse.csr_matrix([[0, 1, 0, 0, 1],
+                                                      [1, 0, 1, 0, 0],
+                                                      [0, 1, 0, 1, 0],
+                                                      [0, 0, 1, 0, 1],
+                                                      [1, 0, 0, 1, 0]])
+
+    assert graph.num_vertices == 5
+    adj_mat_diff = (graph.adj_mat - expected_adjacency_matrix)
+    adj_mat_diff.eliminate_zeros()
+    assert adj_mat_diff.nnz == 0
+
+    # Make sure we can't do something stupid
+    with pytest.raises(ValueError):
+        _ = sgtl.graph.complete_graph(-10)
+    with pytest.raises(ValueError):
+        _ = sgtl.graph.complete_graph(0)
+
+
+def test_star_graph():
+    # Create a star graph
+    n = 5
+    graph = sgtl.graph.star_graph(n)
+    expected_adjacency_matrix = sp.sparse.csr_matrix([[0, 1, 1, 1, 1],
+                                                      [1, 0, 0, 0, 0],
+                                                      [1, 0, 0, 0, 0],
+                                                      [1, 0, 0, 0, 0],
+                                                      [1, 0, 0, 0, 0]])
+
+    assert graph.num_vertices == 5
+    adj_mat_diff = (graph.adj_mat - expected_adjacency_matrix)
+    adj_mat_diff.eliminate_zeros()
+    assert adj_mat_diff.nnz == 0
+
+    # Make sure we can't do something stupid
+    with pytest.raises(ValueError):
+        _ = sgtl.graph.complete_graph(-10)
+    with pytest.raises(ValueError):
+        _ = sgtl.graph.complete_graph(0)
+
+
+def test_path_graph():
+    # Create a star graph
+    n = 5
+    graph = sgtl.graph.path_graph(n)
+    expected_adjacency_matrix = sp.sparse.csr_matrix([[0, 1, 0, 0, 0],
+                                                      [1, 0, 1, 0, 0],
+                                                      [0, 1, 0, 1, 0],
+                                                      [0, 0, 1, 0, 1],
+                                                      [0, 0, 0, 1, 0]])
+
+    assert graph.num_vertices == 5
+    adj_mat_diff = (graph.adj_mat - expected_adjacency_matrix)
+    adj_mat_diff.eliminate_zeros()
+    assert adj_mat_diff.nnz == 0
+
+    # Make sure we can't do something stupid
+    with pytest.raises(ValueError):
+        _ = sgtl.graph.complete_graph(-10)
+    with pytest.raises(ValueError):
+        _ = sgtl.graph.complete_graph(0)
+
+