+++ tests/test_random.py
@@ -1,14 +1,14 @@
 """
-Tests for the sbm module.
+Tests for the random module.
 """
 import scipy.sparse
 from context import sgtl
-import sgtl.sbm
+import sgtl.random
 
 
 def test_num_pos_edges():
     n = 250
-    assert sgtl.sbm._get_num_pos_edges(n, n, True, True, False) == int((n * (n - 1)) / 2) + n
+    assert sgtl.random._get_num_pos_edges(n, n, True, True, False) == int((n * (n - 1)) / 2) + n
 
 
 def test_sbm():
@@ -17,19 +17,19 @@ def test_sbm():
 
     # Generate a graph with this probability matrix, and the following cluster sizes
     cluster_sizes = [100, 50, 20, 100]
-    graph = sgtl.sbm.sbm(cluster_sizes, prob_mat_q, directed=True)
+    graph = sgtl.random.sbm(cluster_sizes, prob_mat_q, directed=True)
     assert graph.num_vertices == sum(cluster_sizes)
 
     # Generate a graph with a fixed number of vertices
     n = 1000
-    graph = sgtl.sbm.sbm_equal_clusters(n, 4, prob_mat_q, directed=True)
+    graph = sgtl.random.sbm_equal_clusters(n, 4, prob_mat_q, directed=True)
     assert graph.num_vertices == n
 
     # Generate a graph with a fixed p and q
     n = 2000
     p = 0.8
     q = 0.2
-    graph = sgtl.sbm.ssbm(n, 4, p, q)
+    graph = sgtl.random.ssbm(n, 4, p, q)
 
     # The adjacency matrix should be symmetric
     sym_diff = (graph.adj_mat - graph.adj_mat.transpose())
@@ -43,3 +43,13 @@ def test_sbm():
     # And in a single cluster should be about
     # (n/4) * (n/4) * p
     assert abs((graph.weight(list(range(500)), list(range(500))) / (int(((n/4) * ((n/4) - 1)) / 2) + (n/4))) - 1) <= 0.1
+
+
+def test_erdos_renyi():
+    # Generate a graph
+    n = 1000
+    graph = sgtl.random.erdos_renyi(n, 0.1)
+
+    # Check that the graph has the expected number of vertices and edges.
+    assert graph.num_vertices == n
+    assert abs((graph.volume(range(1000)) / (int(2 * 0.1 * (n * (n - 1)) / 2) + n)) - 1) <= 0.1